================= gfortran version ================= (1) main.f90 	(a)  print statements are aligned to 132char limit in main.f90 (2) functions.f90 	(a) Commented random number for Hessian-Vector (not currently used) 	(b) Threebar truss with 3 design variables (fct=11) is fixed to fit the length 	(c) Threebar truss with 6 design variables is made dummy (too laborious to fix the col length to 132) 	(d) removed the call to Eulersolve (3) rmsebound.f90 	(a) real*8 for an integer value is changed to integer 	(b) removed unused variables (4) post.f90 	(a) A glitch causing seg fault fixed. usage of explicit bounds i.e. 1:ndim is recommended for xvis variable. 	(b) tecplot changed to point type from block (gfortran behaves differently than ifort) 	(c) ge.10 gt.10 are chaged to 20  (CFD is 20) (5) gatool.f90 	(a) real changed to real*8 (6) sobol.f90 	(a) integer kind=4 changed to kind=8 for compatibility when calling (7) Makefilefullprogram, Makefile 	(a) removed the library Eulersolve.a (need to compile the flow solver with gfortran) 	(b) added -fopenmp flag

+++ gatool.f90
@@ -930,10 +930,10 @@
 
       subroutine INV_MAT_2D(A,AINV,det)
       implicit none
-      real, dimension(2,2), intent(in)  :: A
-      real, dimension(2,2), intent(out) :: AINV
-      real,                 intent(out) :: det
-      real :: a11,a21,a12,a22
+      real*8, dimension(2,2), intent(in)  :: A
+      real*8, dimension(2,2), intent(out) :: AINV
+      real*8,                 intent(out) :: det
+      real*8 :: a11,a21,a12,a22
 
       a11 = A(1,1)
       a21 = A(2,1)
@@ -952,10 +952,10 @@
 !+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
       subroutine INV_MAT_3D(A,AINV,det)
       implicit none
-      real, dimension(3,3), intent(in)  :: A
-      real, dimension(3,3), intent(out) :: AINV
-      real,                 intent(out) :: det
-      real :: a11,a21,a31,a12,a22,a32,a13,a23,a33
+      real*8, dimension(3,3), intent(in)  :: A
+      real*8, dimension(3,3), intent(out) :: AINV
+      real*8,                 intent(out) :: det
+      real*8 :: a11,a21,a31,a12,a22,a32,a13,a23,a33
 
       a11 = A(1,1)
       a21 = A(2,1)
